OpenCL test
Test d’une solution OpenCL

+++ eternityII/datamanager.h
@@ -6,7 +6,7 @@
 #include "lifo.h"
 
 #define NB_FILE_POSSIBILITY 10
-static int max_result = 0;
+static uint16_t max_result = 0;
 
 typedef struct
 {
